+++ changelog.md
@@ -36,7 +36,9 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityTableExportModel.all() added, EntityTableExportPanel all/selected radio button setup improved.
 - EntityApplicationPanel, EntityPanel, EntityEditPanel, EntityTablePanel initialization logging added.
-### is.codion.plugin.flatlaf.intellij
+### is.codion.plugin.flatlaf
+- is.codion.plugin.flatlaf module now service only, look and feels moved to a new is.codion.plugin.flatlaf.lookandfeels module.
+### is.codion.plugin.flatlaf.intellij.themes
 - Solid Cherry, Hard Hacker, Dark Ubuntu, Autumn, Iceberg, Copilot and Pink As Heck IntelliJ themes added.
 
 ## 0.18.63